quick and dirty fix to broken links in lprolog.md
Need to figure out how to properly link to a folder's contents.

Things that definitely don't work:

* raw link to folder
* adding a .htaccess with Options +Indexes

Things that may work: using the Github REST api. Example:

https://api.github.com/repos/distributed-assertions/distributed-assertions.github.io/contents/assets/lProlog?ref=gh-pages

We will then need to add some code to parse and extract a directory listing from
it. Would be easy if there was a way to embed a <script> inside mkdocs (needs
further investigation).